Migrate account list to new user model
