+++ src/main/java/com/owncloud/android/ui/activity/UserInfoActivity.java
@@ -3,11 +3,11 @@
  *
  * @author Mario Danic
  * @author Andy Scherzinger
- * @author Chris Narkiewicz
+ * @author Chris Narkiewicz  <hello@ezaquarii.com>
  * Copyright (C) 2017 Mario Danic
  * Copyright (C) 2017 Andy Scherzinger
  * Copyright (C) 2017 Nextcloud GmbH.
- * Copyright (C) 2019 Chris Narkiewicz <hello@ezaquarii.com>
+ * Copyright (C) 2020 Chris Narkiewicz <hello@ezaquarii.com>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published by
@@ -50,8 +50,10 @@ import com.bumptech.glide.request.animation.GlideAnimation;
 import com.bumptech.glide.request.target.SimpleTarget;
 import com.evernote.android.job.JobRequest;
 import com.evernote.android.job.util.support.PersistableBundleCompat;
+import com.nextcloud.client.account.User;
 import com.nextcloud.client.di.Injectable;
 import com.nextcloud.client.preferences.AppPreferences;
+import com.nextcloud.java.util.Optional;
 import com.owncloud.android.R;
 import com.owncloud.android.jobs.AccountRemovalJob;
 import com.owncloud.android.lib.common.UserInfo;
@@ -116,7 +118,7 @@ public class UserInfoActivity extends FileActivity implements Injectable {
     private Unbinder unbinder;
 
     private UserInfo userInfo;
-    private Account account;
+    private User user;
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -124,7 +126,14 @@ public class UserInfoActivity extends FileActivity implements Injectable {
         super.onCreate(savedInstanceState);
         Bundle bundle = getIntent().getExtras();
 
-        account = Parcels.unwrap(bundle.getParcelable(KEY_ACCOUNT));
+        final Account account = Parcels.unwrap(bundle.getParcelable(KEY_ACCOUNT));
+        Optional<User> optionalUser = accountManager.getUser(account != null ? account.name : "");
+        if(!optionalUser.isPresent()) {
+            finish();
+            return;
+        } else {
+            user = optionalUser.get();
+        }
 
         if (savedInstanceState != null && savedInstanceState.containsKey(KEY_USER_DATA)) {
             userInfo = Parcels.unwrap(savedInstanceState.getParcelable(KEY_USER_DATA));
@@ -169,7 +178,7 @@ public class UserInfoActivity extends FileActivity implements Injectable {
                 onBackPressed();
                 break;
             case R.id.delete_account:
-                openAccountRemovalConfirmationDialog(account, getSupportFragmentManager());
+                openAccountRemovalConfirmationDialog(user, getSupportFragmentManager());
                 break;
             default:
                 retval = super.onOptionsItemSelected(item);
@@ -209,13 +218,13 @@ public class UserInfoActivity extends FileActivity implements Injectable {
     }
 
     private void setHeaderImage() {
-        if (getStorageManager().getCapability(account.name).getServerBackground() != null) {
+        if (getStorageManager().getCapability(user.getAccountName()).getServerBackground() != null) {
             ViewGroup appBar = findViewById(R.id.appbar);
 
             if (appBar != null) {
                 ImageView backgroundImageView = appBar.findViewById(R.id.drawer_header_background);
 
-                String background = getStorageManager().getCapability(account.name).getServerBackground();
+                String background = getStorageManager().getCapability(user.getAccountName()).getServerBackground();
                 int primaryColor = ThemeUtils.primaryColor(getAccount(), false, this);
 
                 if (URLUtil.isValidUrl(background)) {
@@ -253,11 +262,11 @@ public class UserInfoActivity extends FileActivity implements Injectable {
     }
 
     private void populateUserInfoUi(UserInfo userInfo) {
-        userName.setText(account.name);
-        avatar.setTag(account.name);
-        DisplayUtils.setAvatar(account, this, mCurrentAccountAvatarRadiusDimension, getResources(), avatar, this);
+        userName.setText(user.getAccountName());
+        avatar.setTag(user.getAccountName());
+        DisplayUtils.setAvatar(user, this, mCurrentAccountAvatarRadiusDimension, getResources(), avatar, this);
 
-        int tint = ThemeUtils.primaryColor(account, true, this);
+        int tint = ThemeUtils.primaryColor(user.toPlatformAccount(), true, this);
 
         if (!TextUtils.isEmpty(userInfo.getDisplayName())) {
             fullName.setText(userInfo.getDisplayName());
@@ -299,9 +308,9 @@ public class UserInfoActivity extends FileActivity implements Injectable {
         }
     }
 
-    public static void openAccountRemovalConfirmationDialog(Account account, FragmentManager fragmentManager) {
+    public static void openAccountRemovalConfirmationDialog(User user, FragmentManager fragmentManager) {
         UserInfoActivity.AccountRemovalConfirmationDialog dialog =
-            UserInfoActivity.AccountRemovalConfirmationDialog.newInstance(account);
+            UserInfoActivity.AccountRemovalConfirmationDialog.newInstance(user);
         dialog.show(fragmentManager, "dialog");
     }
 
@@ -309,9 +318,9 @@ public class UserInfoActivity extends FileActivity implements Injectable {
 
         private Account account;
 
-        public static UserInfoActivity.AccountRemovalConfirmationDialog newInstance(Account account) {
+        public static UserInfoActivity.AccountRemovalConfirmationDialog newInstance(User user) {
             Bundle bundle = new Bundle();
-            bundle.putParcelable(KEY_ACCOUNT, account);
+            bundle.putParcelable(KEY_ACCOUNT, user.toPlatformAccount());
 
             UserInfoActivity.AccountRemovalConfirmationDialog dialog = new
                     UserInfoActivity.AccountRemovalConfirmationDialog();
@@ -366,7 +375,7 @@ public class UserInfoActivity extends FileActivity implements Injectable {
     private void fetchAndSetData() {
         Thread t = new Thread(() -> {
             RemoteOperation getRemoteUserInfoOperation = new GetUserInfoRemoteOperation();
-            RemoteOperationResult result = getRemoteUserInfoOperation.execute(account, this);
+            RemoteOperationResult result = getRemoteUserInfoOperation.execute(user.toPlatformAccount(), this);
 
             if (result.isSuccess() && result.getData() != null) {
                 userInfo = (UserInfo) result.getData().get(0);
